requirements.txt
beautifulsoup4
camoufox
curl_cffi
patchright
playwright
python-dotenv
quart
requests
rich
